kazani/admin.py
from django.contrib import admin
from unfold.admin import ModelAdmin

from .models import User


@admin.register(User)
class UserAdmin(ModelAdmin):
    readonly_fields = ("date_joined", "last_login", "password")
