pyproject.toml
[tool.poetry]
name = "kaz-cms"
version = "0.1.0"
description = ""
authors = ["Ohin Taylor <kazani@kazani.dev>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = {extras = ["argon2"], version = "^5.0.7"}
uvicorn = "^0.30.1"
django-taggit = "^5.0.1"
base32-crockford = "^0.3.0"
emoji = "^2.12.1"
django-compressor = "^4.5"
django-unfold = "^0.34.0"
marko = "^2.1.2"
django-minify-html = "^1.10.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"
black = "^24.4.2"
pylint = "^3.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
